Added Auth for Deletes and Updates
Ironed out edge cases for Images, Posts, and Comments
Routes should be mostly secure and returns proper error messages

+++ MarketPlaceBackend/MarketPlaceBackend/Controllers/CommentController.cs
@@ -1,8 +1,10 @@
 using MarketPlaceBackend.Data;
 using MarketPlaceBackend.DTOs;
 using MarketPlaceBackend.Models;
+using Microsoft.AspNetCore.Authorization;
 using Microsoft.AspNetCore.Mvc;
 using Microsoft.EntityFrameworkCore;
+using System.Security.Claims;
 
 namespace MarketPlaceBackend.Controllers;
 
@@ -20,12 +22,18 @@ public class CommentController : ControllerBase
     }
 
     [HttpPost]
+    [Authorize]
     public async Task<IActionResult> CreateNewComment(CommentDTO commentDto)
     {
-        Comments comment = new Comments()
+        var userId = User.FindFirstValue(ClaimTypes.NameIdentifier);
+
+        if (userId == null)
+            return BadRequest("Cannot Validate User");
+        
+        Comments comment = new Comments
         {
             PostId = commentDto.PostId,
-            UserId = commentDto.UserId,
+            UserId = userId,
             Content = commentDto.Content,
             CreatedAt = DateTime.UtcNow
         };
@@ -33,7 +41,7 @@ public class CommentController : ControllerBase
         await _db.Comments.AddAsync(comment);
         await _db.SaveChangesAsync();
 
-        _logger.LogEvent($"User {commentDto.UserId} commented on Post {commentDto.PostId}");
+        _logger.LogEvent($"User {userId} commented on Post {commentDto.PostId}");
         
         return NoContent();
     }
@@ -54,18 +62,30 @@ public class CommentController : ControllerBase
                     UpdatedAt = c.UpdatedAt
                 }
             ).ToList();
+ 
+        // decide on return empty array or a NotFound();
 
         return Ok(comments);
     }
 
     [HttpPut]
+    [Authorize]
     public IActionResult UpdateComment(int commentId, UpdatedCommentDTOs commentDto)
     {
+        var userId = User.FindFirstValue(ClaimTypes.NameIdentifier);
+
+        if (userId == null)
+            return BadRequest("Cannot Validate User");
+        
         var comment = _db.Comments
             .FirstOrDefault(c => c.Id == commentId);
-
+        
         if (comment == null)
             return NotFound();
+        
+        // The logged in user and the comments owner don't match
+        if (comment.UserId != userId)
+            return BadRequest("UserId Mismatch");
 
         comment.Content = commentDto.Content;
 
@@ -77,14 +97,24 @@ public class CommentController : ControllerBase
     }
 
     [HttpDelete]
+    [Authorize]
     public async Task<IActionResult> DeleteComment(int commentId)
     {
+        var userId = User.FindFirstValue(ClaimTypes.NameIdentifier);
+
+        if (userId == null)
+            return BadRequest("Cannot Validate User");
+        
         var comment = await _db.Comments.FindAsync(commentId);
 
         if (comment == null)
             return NotFound();
+        
+        if (comment.UserId != userId)
+            return BadRequest("UserId Mismatch");
 
         _db.Comments.Remove(comment);
+        
         await _db.SaveChangesAsync();
         
         _logger.LogEvent($"User {comment.UserId} deleted Comment {comment.Id}");
